+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) since v1.0.0.
 
 ## [Unreleased]
+### Fixed
+- On click on the plus icon, a new field is added on the `TextFieldList` component
 
 ## [v0.0.9] - 2019-08-28
 ### Added